requirements.txt
discord.py
requests
asyncio
json